Move libgit2 settings into libgit2.c
